+++ src/channels/middleware/scenario-middleware.js
@@ -1,14 +1,14 @@
 var RunChannel = require('./run-middleware');
+var CustomRunChannel = require('./custom-run-middleware');
 
 var prefix = require('./middleware-utils').prefix;
-
 var mapWithPrefix = require('./middleware-utils').mapWithPrefix;
-var CustomRunChannel = require('./custom-run-middleware');
 
 var Middleware = require('./general-middleware');
 module.exports = function (config, notifier) {
     var defaults = {
         serviceOptions: {},
+        channelOptions: {},
     };
     var opts = $.extend(true, {}, defaults, config);
 
@@ -26,20 +26,23 @@ module.exports = function (config, notifier) {
     var baselineRunChannel = new RunChannel( 
         $.extend(true, {}, {
             serviceOptions: baselinePromise,
-            meta: {
-                readOnly: true
-            },
-            variables: {
-                readOnly: true
+            channelOptions: {
+                meta: {
+                    readOnly: true
+                },
+                variables: {
+                    readOnly: true
+                }
             }
         }, opts.defaults, opts.baseline)
     , notifyWithPrefix.bind(null, 'baseline:'));
-    var currentRunChannel = new RunChannel($.extend(true, {}, {
+
+    var runOptions = $.extend(true, {}, {
         serviceOptions: currentRunPromise,
-    }, opts.defaults, opts.current), notifyWithPrefix.bind(null, 'current:'));
-    var defaultRunChannel = new RunChannel($.extend(true, {}, {
-        serviceOptions: currentRunPromise,
-    }, opts.defaults, opts.current), notifyWithPrefix.bind(null, ''));
+    }, opts.defaults, opts.current);
+
+    var currentRunChannel = new RunChannel(runOptions, notifyWithPrefix.bind(null, 'current:'));
+    var defaultRunChannel = new RunChannel(runOptions, notifyWithPrefix.bind(null, ''));
     var customRunChannel = new CustomRunChannel(opts.serviceOptions.run, notifyWithPrefix);
 
     var handlers = [
@@ -49,7 +52,7 @@ module.exports = function (config, notifier) {
         $.extend(defaultRunChannel, { name: 'default', match: prefix('') }),
     ];
     
-    var middleware = new Middleware(handlers, config, notifier);
+    var middleware = new Middleware(handlers, notifier);
     middleware.scenarioManager = sm;
     return middleware;
 };